+++ examples/src/main/kotlin/examples/HelloWorldShadowWithPlane.kt
@@ -24,7 +24,7 @@ class HelloWorldShadowWithPlane {
 
     init {
         scene = Scene()
-        camera = PerspectiveCamera(75, window.innerWidth / window.innerHeight, 0.1, 1000)
+        camera = PerspectiveCamera(75, (window.innerWidth / window.innerHeight).toDouble(), 0.1, 1000)
 
         val params = WebGLRendererParams()
         params.antialias = true
